+++ spec/number-display-spec.js
@@ -165,6 +165,37 @@ describe('dc.numberDisplay', function () {
             d3.select('#number-display-test-section').remove();
         });
     });
+    describe('with group with multiple values', function () {
+        var group, chart;
+        beforeEach(function () {
+            countryDimension.filterAll();
+            group = countryDimension.group().reduceSum(function (d) { return +d.value; });
+            var id = 'empty-div';
+            appendChartID(id);
+            chart = buildChart('#' + id);
+            chart
+                .group(group)
+                .valueAccessor(function (kv) { return kv.value; })
+                .render();
+            d3.timer.flush();
+        });
+
+        it('should show the largest value', function () {
+            expect(chart.select('span.number-display').text()).toEqual('341');
+        });
+
+        describe('with reversed ordering', function () {
+            beforeEach(function () {
+                chart.ordering(function (kv) { return -kv.value; })
+                    .render();
+                d3.timer.flush();
+            });
+            it('should show the smallest value', function () {
+                expect(chart.select('span.number-display').text()).toEqual('77.0');
+            });
+        });
+
+    });
     describe('Infinity', function () {
         var chart;
         beforeEach(function () {